Use Hugo Pipes to convert SCSS files to CSS
Hugp Pipes makes it possible to easily convert and minify SCSS files
located in the 'assets' directory. With this change, I don't need to run
the SCSS preprocessor besides hugo server. Furthermore, this should make
things much easier in the CI pipeline.

I moved all the .scss files from the 'static' to the 'assets' directory.
Next I replaced all the references to .css files in the templates with a
partial that generates the appropriate <link>. The partial handles the
conversion to CSS, minification, and fingerprinting.

For this feature to work, the extended version of Hugo is required.