Added template for Atom feed
Hugo only provides built-in support for RSS. But RSS does not offer the
possibility to include content other than plain text. Therefore, users
have to visit the website to read the content, but cannot do so from
their feed reader.

With Atom, it is possible to include the HTML content.

A current problem is that the HTML is not minified and the atom.xml
looks messy.