Use Hugo Pipes to convert SCSS files to CSS (#2)
Hugo Pipes makes it possible to easily convert and minify SCSS files
located in the _assets_ directory. With this commit, I don't need to run
the SCSS preprocessor besides `hugo server` anymore. Furthermore, this should make
things much easier in the CI pipeline.

I moved all the .scss files from the _static_ to the _assets_ directory.
Next, I replaced all the references to .css files in the templates with a
partial that generates the correct `<link>` tag. The partial handles the
conversion to CSS, minification, and fingerprinting.

For this feature to work, the extended version of Hugo is required.