Add Atom feed (#3)
Hugo only provides built-in support for RSS. But RSS does not offer the
possibility to include content other than plain text. Therefore, users
have to visit the website to read the content, but cannot do so from
their feed reader.

With Atom, it is possible to include the HTML content.

I created a template for the Atom feed and configured the site to not generate any RSS files, but only one index.atom.xml for the homepage. (I don't know how I can use atom.xml as name for this file.) The template is validated and should conform to RFC 4287.

The footer now contains a link to the Atom XML file.