Create distinct configs for environments (#5)
I want to be able to adjust configuration based on the environment, e.g. the baseURL or the number of articles per page.

Therefore, I created a config directory and moved the config file to _default. I also created a production directory which contains the config for this environment. It overrides the settings in the _default directory.

Because I still can't read TOML, I also converted the files to YAML.