Use Dart Sass (#8)
Hugo Pipes uses LibSass, but it is not fully compatible with the latest Sass specification and does lack features like the new module system. I used Hugo Pipes to convert SCSS to CSS and to add Subresource Integrity (SRI).

Dart Sass is now used to process the .scss files. It runs in the build pipeline after Hugo. Hugo moves all files and folders in the static directory to the output directory. Dart Sass processes the content of the [outputDir]/style directory and removes all .scss files afterwards.

Currently, all .scss files are processed, but this does only take up a few more bytes on the server — the browser will only download the .css files it needs.

I removed SRI. Honestly, I only used it because it was so easy to add. Because I'm currently not using a CDN and SRI can sometimes lead to problems (it was with caching, if I remember correctly), I will not add it again in the near future.