Use @use and @forward Sass rules
The new Sass module system replaces the @import statements.

I also prefixed all .scss files which should not be converted to CSS
(because they are only @used or @forwarded) with _ and moved them to the
'components' directory.

I cannot write "@use 'variables' as *", because then all variables will
be imported multiple times in files like article.scss, which would cause
an error. Therefore I have to use a namespace.