Made #articleContent a CSS class
I made the mistake to use ids instead of classes in my stylesheets. The
problem is that ids have a too high specificity, which can confound the
intuitive cascade. For example, '#articleContent a' has a higher
specificity than '.socialMediaLink'.

With .articleContent as class, I can use the well-known CSS hack to
repeat classes in order to increase specificity. I have done this in
_reference.scss.